% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range.R
\name{date_range}
\alias{date_range}
\title{Enumerate a range of days starting from one time to another time.}
\usage{
date_range(beginning, end)
}
\arguments{
\item{beginning}{character. A relative time character expression (e.g., "4 days ago").}

\item{end}{character. A relative time character expression (e.g., "now").}
}
\description{
Enumerate a range of days starting from one time to another time.
}
