% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strdate.R
\docType{data}
\name{strdate}
\alias{strdate}
\title{Convert a relative time character expression into an absolute date expression.}
\format{An object of class \code{validated_function} (inherits from \code{logical}) of length 1.}
\usage{
strdate
}
\arguments{
\item{time}{character. The time character expression to convert.}
}
\description{
Convert a relative time character expression into an absolute date expression.
}
\keyword{datasets}

