% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strdate.R
\name{strdate}
\alias{strdate}
\title{Convert a relative time character expression into an absolute date expression.}
\usage{
strdate(time, relative_to = Sys.time())
}
\arguments{
\item{time}{character. The time character expression to convert.}

\item{relative_to}{POSIXt. Relative to what time are we converting?}
}
\description{
Convert a relative time character expression into an absolute date expression.
}
